fix(edit): certify bounded cross-platform constraints
