+++ tests/unit/cli/test_constraint_check_command.py
@@ -3,17 +3,13 @@
 from __future__ import annotations
 
 import json
-import os
 import sqlite3
 from pathlib import Path
 from types import SimpleNamespace
-from unittest.mock import AsyncMock, MagicMock, patch
-
-import pytest
+from unittest.mock import patch
 
 from tree_sitter_analyzer.cli.commands.constraint_check_command import (
     _compute_verdict,
-    _evaluate_with_explicit_file,
     _exit_code_for,
     _failure_envelope,
     _filter_violations,
@@ -21,11 +17,8 @@ from tree_sitter_analyzer.cli.commands.constraint_check_command import (
     _load_explicit,
     _print_result,
     _resolve_output_format,
-    _run_and_persist,
-    _run_tool,
     _violations_ddl,
     get_default_project_root,
-    run_check_constraints,
 )
 
 # Module-level patch targets
@@ -408,501 +401,3 @@ class TestLoadExplicit:
             _load_explicit(yaml_file)
 
         assert file_contents[0] == content
-
-
-# ---------------------------------------------------------------------------
-# _run_and_persist
-# ---------------------------------------------------------------------------
-
-
-class TestRunAndPersist:
-    def _empty_db(self, tmp_path: Path) -> Path:
-        db = tmp_path / "index.db"
-        sqlite3.connect(str(db)).close()
-        return db
-
-    def _db_with_edges(self, tmp_path: Path) -> Path:
-        from tree_sitter_analyzer.graph.edge_store import EDGE_STORE_SCHEMA
-
-        db = tmp_path / "index.db"
-        conn = sqlite3.connect(str(db))
-        # B1.3: the edge-count gate counts CALLS rows in the unified ``edges``
-        # table (ast_call_edges was dropped).
-        conn.executescript(EDGE_STORE_SCHEMA)
-        conn.execute(
-            "INSERT INTO edges (source_node_id, target_node_id, kind) "
-            "VALUES ('a.py:f:1', 'b.py:g:1', 'calls')"
-        )
-        conn.commit()
-        conn.close()
-        return db
-
-    def test_no_call_edges_table_returns_empty(self, tmp_path):
-        db = self._empty_db(tmp_path)
-        violations, edge_count = _run_and_persist(db, [])
-        assert violations == []
-        assert edge_count == 0
-
-    @pytest.mark.slow_ok  # Windows xdist-load budget exemption; test logic is trivial, no perf claim (#976)
-    def test_empty_call_edges_table_returns_empty(self, tmp_path):
-        db = tmp_path / "index.db"
-        conn = sqlite3.connect(str(db))
-        conn.execute("CREATE TABLE ast_call_edges (id INTEGER PRIMARY KEY)")
-        conn.commit()
-        conn.close()
-        violations, edge_count = _run_and_persist(db, [])
-        assert violations == []
-        assert edge_count == 0
-
-    def test_evaluate_exception_degrades_gracefully(self, tmp_path):
-        db = self._db_with_edges(tmp_path)
-        with patch(_EVALUATE, side_effect=RuntimeError("boom")):
-            violations, edge_count = _run_and_persist(db, [])
-        assert violations == []
-        assert edge_count == 1
-
-    def test_violations_persisted_to_db(self, tmp_path):
-        db = self._db_with_edges(tmp_path)
-        v = _v(detected_at=12345)
-        with patch(_EVALUATE, return_value=[v]):
-            violations, _ = _run_and_persist(db, ["c"])
-        assert len(violations) == 1
-        conn = sqlite3.connect(str(db))
-        rows = conn.execute("SELECT rule_id FROM ast_constraint_violations").fetchall()
-        conn.close()
-        assert rows == [("R1",)]
-
-    def test_returns_edge_count_from_db(self, tmp_path):
-        db = self._db_with_edges(tmp_path)
-        with patch(_EVALUATE, return_value=[]):
-            _, edge_count = _run_and_persist(db, [])
-        assert edge_count == 1
-
-    def test_read_only_evaluation_returns_rows_without_creating_cache_table(
-        self, tmp_path
-    ):
-        db = self._db_with_edges(tmp_path)
-        violation = _v()
-
-        with patch(_EVALUATE, return_value=[violation]):
-            result = _run_and_persist(db, ["c"], persist=False)
-
-        conn = sqlite3.connect(str(db))
-        tables = conn.execute(
-            "SELECT name FROM sqlite_master "
-            "WHERE type = 'table' AND name = 'ast_constraint_violations'"
-        ).fetchall()
-        conn.close()
-        assert result == ([violation], 1)
-        assert tables == []
-
-    def test_read_only_missing_edges_table_reraises_operational_error(self, tmp_path):
-        db = self._empty_db(tmp_path)
-
-        with pytest.raises(sqlite3.OperationalError, match="no such table: edges"):
-            _run_and_persist(db, [], persist=False)
-
-    def test_read_only_evaluator_exception_is_not_degraded(self, tmp_path):
-        db = self._db_with_edges(tmp_path)
-
-        with patch(_EVALUATE, side_effect=RuntimeError("evaluation failed")):
-            with pytest.raises(RuntimeError, match="^evaluation failed$"):
-                _run_and_persist(db, [], persist=False)
-
-    def test_violations_table_cleared_before_insert(self, tmp_path):
-        db = self._db_with_edges(tmp_path)
-        # Pre-populate violations table with a stale row
-        conn = sqlite3.connect(str(db))
-        conn.execute(_violations_ddl())
-        conn.execute(
-            """INSERT INTO ast_constraint_violations
-               VALUES ('OLD', 'f.py', 'fn', 1, 'bar', 'g.py', 'warn', 0)"""
-        )
-        conn.commit()
-        conn.close()
-
-        new_v = _v(rule_id="NEW", detected_at=1)
-        with patch(_EVALUATE, return_value=[new_v]):
-            _run_and_persist(db, ["c"])
-
-        conn = sqlite3.connect(str(db))
-        rows = conn.execute("SELECT rule_id FROM ast_constraint_violations").fetchall()
-        conn.close()
-        # OLD row must be gone; only NEW should be present
-        rule_ids = [r[0] for r in rows]
-        assert "OLD" not in rule_ids
-        assert "NEW" in rule_ids
-
-    def test_duplicate_pk_violations_do_not_crash_persist(self, tmp_path):
-        """Regression for #544: two violations with the same PK must not crash.
-
-        If ``evaluate()`` returns two ``Violation`` objects that share the
-        same ``(rule_id, caller_file, caller_line, callee_name)`` PRIMARY
-        KEY (e.g., one call site resolved to two ``callee_file`` targets),
-        the old ``executemany`` would raise
-        ``UNIQUE constraint failed: ast_constraint_violations.rule_id, ...``.
-
-        After the fix the persist path must succeed and write exactly 1 row
-        (the dedup is in ``evaluate()``, so ``_run_and_persist`` receives a
-        clean list — this test verifies the full stack from mock to DB).
-        """
-        db = self._db_with_edges(tmp_path)
-        # Two violations with identical PK but different callee_file.
-        dup_v1 = _v(
-            rule_id="R1",
-            caller_file="a.py",
-            caller_line=10,
-            callee_name="bar",
-            callee_file="b.py",
-            detected_at=1,
-        )
-        dup_v2 = _v(
-            rule_id="R1",
-            caller_file="a.py",
-            caller_line=10,
-            callee_name="bar",
-            callee_file="c.py",
-            detected_at=1,
-        )
-
-        # We intentionally bypass the real evaluate() and inject the two
-        # duplicates directly to test the persist layer in isolation.
-        with patch(_EVALUATE, return_value=[dup_v1, dup_v2]):
-            # Must NOT raise sqlite3.IntegrityError.
-            violations, edge_count = _run_and_persist(db, ["c"])
-
-        conn = sqlite3.connect(str(db))
-        rows = conn.execute(
-            "SELECT rule_id, caller_file, caller_line, callee_name "
-            "FROM ast_constraint_violations"
-        ).fetchall()
-        conn.close()
-        # Exactly 1 row persisted (PK is unique); the constraint did not crash.
-        assert len(rows) == 1, (
-            f"Expected exactly 1 persisted row after dedup, got {len(rows)}: {rows}"
-        )
-        assert rows[0] == ("R1", "a.py", 10, "bar")
-
-
-# ---------------------------------------------------------------------------
-# _run_tool
-# ---------------------------------------------------------------------------
-
-
-class TestRunTool:
-    def test_builds_tool_and_returns_execute_coroutine(self, tmp_path):
-        import asyncio
-
-        mock_tool = MagicMock()
-        mock_tool.execute = AsyncMock(return_value={"success": True})
-
-        with patch(_CCT_CLS, return_value=mock_tool):
-            result = asyncio.run(_run_tool(str(tmp_path), "warn", "", "json"))
-
-        assert result == {"success": True}
-        mock_tool.execute.assert_called_once_with(
-            {"path_filter": "", "severity_min": "warn", "output_format": "json"}
-        )
-
-    def test_passes_path_filter_and_severity(self, tmp_path):
-        import asyncio
-
-        mock_tool = MagicMock()
-        mock_tool.execute = AsyncMock(return_value={"success": False})
-
-        with patch(_CCT_CLS, return_value=mock_tool):
-            asyncio.run(_run_tool(str(tmp_path), "error", "src/*", "toon"))
-
-        called_payload = mock_tool.execute.call_args[0][0]
-        assert called_payload["severity_min"] == "error"
-        assert called_payload["path_filter"] == "src/*"
-        assert called_payload["output_format"] == "toon"
-
-    def test_read_only_omits_persistence_from_tool_execution(self, tmp_path):
-        import asyncio
-
-        mock_tool = MagicMock()
-        mock_tool.execute = AsyncMock(return_value={"success": True})
-
-        with patch(_CCT_CLS, return_value=mock_tool):
-            asyncio.run(_run_tool(str(tmp_path), "warn", "", "json", persist=False))
-
-        mock_tool.execute.assert_awaited_once_with(
-            {
-                "path_filter": "",
-                "severity_min": "warn",
-                "output_format": "json",
-                "persist": False,
-            }
-        )
-
-
-# ---------------------------------------------------------------------------
-# _evaluate_with_explicit_file
-# ---------------------------------------------------------------------------
-
-
-class TestEvaluateWithExplicitFile:
-    def _call(
-        self,
-        tmp_path: Path,
-        constraint_file: str,
-        *,
-        severity_min: str = "warn",
-        path_filter: str = "",
-        output_format: str = "json",
-        persist: bool = True,
-    ) -> dict:
-        return _evaluate_with_explicit_file(
-            project_root=str(tmp_path),
-            constraint_file=constraint_file,
-            severity_min=severity_min,
-            path_filter=path_filter,
-            output_format=output_format,
-            persist=persist,
-        )
-
-    def test_file_not_found_returns_failure(self, tmp_path):
-        with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-            result = self._call(tmp_path, str(tmp_path / "missing.yml"))
-        assert result["success"] is False
-        assert "not found" in result["error"]
-
-    def test_parse_error_returns_failure(self, tmp_path):
-        from tree_sitter_analyzer.constraints.parser import ConstraintParseError
-
-        yaml_file = tmp_path / "constraints.yml"
-        yaml_file.write_text("")
-        with patch(_LOAD_EXPLICIT, side_effect=ConstraintParseError("bad")):
-            with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-                result = self._call(tmp_path, str(yaml_file))
-        assert result["success"] is False
-        assert "parse error" in result["error"]
-
-    def test_no_db_returns_safe_with_note(self, tmp_path):
-        yaml_file = tmp_path / "constraints.yml"
-        yaml_file.write_text("")
-        with patch(_LOAD_EXPLICIT, return_value=[]):
-            with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-                result = self._call(tmp_path, str(yaml_file))
-        assert result["verdict"] == "SAFE"
-        assert "note" in result
-        assert result["evaluated_edge_count"] == 0
-
-    def test_with_db_no_violations_returns_safe(self, tmp_path):
-        yaml_file = tmp_path / "constraints.yml"
-        yaml_file.write_text("")
-        db_dir = tmp_path / ".ast-cache"
-        db_dir.mkdir()
-        sqlite3.connect(str(db_dir / "index.db")).close()
-
-        with patch(_LOAD_EXPLICIT, return_value=[]):
-            with patch(_RUN_AND_PERSIST, return_value=([], 5)):
-                with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-                    result = self._call(tmp_path, str(yaml_file))
-        assert result["verdict"] == "SAFE"
-        assert result["success"] is True
-        assert result["evaluated_edge_count"] == 5
-
-    def test_read_only_missing_edges_fails_closed_with_nonzero_exit(self, tmp_path):
-        # PR #1254 review 3766246590: an absent edge capability is not SAFE.
-        yaml_file = tmp_path / "constraints.yml"
-        yaml_file.write_text("")
-        db_dir = tmp_path / ".ast-cache"
-        db_dir.mkdir()
-        sqlite3.connect(str(db_dir / "index.db")).close()
-
-        with patch(_LOAD_EXPLICIT, return_value=[]):
-            with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-                result = self._call(tmp_path, str(yaml_file), persist=False)
-
-        expected_error = (
-            "CORRUPT_INDEX" if os.name == "posix" else "SECURE_FD_SNAPSHOT_UNSUPPORTED"
-        )
-        assert (
-            result["success"],
-            result["verdict"],
-            result["error_code"],
-            result["error"],
-            result["violations"],
-            result["rule_count"],
-            _exit_code_for(result),
-        ) == (
-            False,
-            "ERROR",
-            "CONSTRAINT_INDEX_UNKNOWN",
-            expected_error,
-            [],
-            0,
-            1,
-        )
-
-    def test_read_only_corrupt_edges_fails_closed_with_nonzero_exit(self, tmp_path):
-        # PR #1254 review 3766246590: evaluator database failures are not SAFE.
-        yaml_file = tmp_path / "constraints.yml"
-        yaml_file.write_text("")
-        db_dir = tmp_path / ".ast-cache"
-        db_dir.mkdir()
-        conn = sqlite3.connect(str(db_dir / "index.db"))
-        conn.execute("CREATE TABLE edges(kind TEXT)")
-        conn.execute("INSERT INTO edges VALUES ('calls')")
-        conn.commit()
-        conn.close()
-
-        with patch(_LOAD_EXPLICIT, return_value=[]):
-            with patch(_EVALUATE, side_effect=sqlite3.DatabaseError("CORRUPT_INDEX")):
-                with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-                    result = self._call(tmp_path, str(yaml_file), persist=False)
-
-        expected_error = (
-            "CORRUPT_INDEX" if os.name == "posix" else "SECURE_FD_SNAPSHOT_UNSUPPORTED"
-        )
-        assert (
-            result["success"],
-            result["verdict"],
-            result["error_code"],
-            result["error"],
-            result["violations"],
-            result["rule_count"],
-            _exit_code_for(result),
-        ) == (
-            False,
-            "ERROR",
-            "CONSTRAINT_INDEX_UNKNOWN",
-            expected_error,
-            [],
-            0,
-            1,
-        )
-
-    def test_constraint_file_path_included_in_result(self, tmp_path):
-        yaml_file = tmp_path / "constraints.yml"
-        yaml_file.write_text("")
-        db_dir = tmp_path / ".ast-cache"
-        db_dir.mkdir()
-        sqlite3.connect(str(db_dir / "index.db")).close()
-
-        with patch(_LOAD_EXPLICIT, return_value=[]):
-            with patch(_RUN_AND_PERSIST, return_value=([], 0)):
-                with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-                    result = self._call(tmp_path, str(yaml_file))
-        assert "constraint_file" in result
-
-    def test_with_db_and_error_violations_returns_unsafe(self, tmp_path):
-        yaml_file = tmp_path / "constraints.yml"
-        yaml_file.write_text("")
-        db_dir = tmp_path / ".ast-cache"
-        db_dir.mkdir()
-        sqlite3.connect(str(db_dir / "index.db")).close()
-
-        error_v = _v(severity="error")
-        with patch(_LOAD_EXPLICIT, return_value=[]):
-            with patch(_RUN_AND_PERSIST, return_value=([error_v], 3)):
-                with patch(_APPLY_TOON, side_effect=lambda p, fmt: p):
-                    result = self._call(tmp_path, str(yaml_file))
-        assert result["verdict"] == "UNSAFE"
-
-
-# ---------------------------------------------------------------------------
-# run_check_constraints — main dispatcher
-# ---------------------------------------------------------------------------
-
-
-def _ns(**kwargs: object) -> SimpleNamespace:
-    return SimpleNamespace(
-        severity_min=kwargs.get("severity_min", "warn"),
-        constraint_path_filter=kwargs.get("constraint_path_filter", ""),
-        constraint_file=kwargs.get("constraint_file", None),
-        constraints_read_only=kwargs.get("constraints_read_only", False),
-    )
-
-
-class TestRunCheckConstraints:
-    def test_with_constraint_file_routes_to_evaluate_explicit(self, tmp_path):
-        args = _ns(constraint_file="/some/path.yml")
-        safe_result = {"success": True, "verdict": "SAFE"}
-        with patch(_RESOLVE_OFMT, return_value="json"):
-            with patch(_EVAL_EXPLICIT, return_value=safe_result) as mock_eval:
-                with patch(_PRINT_RESULT):
-                    code = run_check_constraints(args, str(tmp_path))
-        mock_eval.assert_called_once()
-        assert code == 0
-
-    def test_without_constraint_file_calls_asyncio_run(self, tmp_path):
-        args = _ns()
-        safe_result = {"success": True, "verdict": "SAFE"}
-        with patch(_RESOLVE_OFMT, return_value="json"):
-            with patch(_ASYNCIO_RUN, return_value=safe_result) as mock_run:
-                with patch(_PRINT_RESULT):
-                    code = run_check_constraints(args, str(tmp_path))
-        mock_run.assert_called_once()
-        assert code == 0
-
-    def test_caution_verdict_returns_exit_2(self, tmp_path):
-        args = _ns()
-        caution_result = {"success": True, "verdict": "CAUTION"}
-        with patch(_RESOLVE_OFMT, return_value="json"):
-            with patch(_ASYNCIO_RUN, return_value=caution_result):
-                with patch(_PRINT_RESULT):
-                    code = run_check_constraints(args, str(tmp_path))
-        assert code == 2
-
-    def test_failure_result_returns_exit_1(self, tmp_path):
-        args = _ns()
-        fail_result = {"success": False, "verdict": "UNSAFE"}
-        with patch(_RESOLVE_OFMT, return_value="json"):
-            with patch(_ASYNCIO_RUN, return_value=fail_result):
-                with patch(_PRINT_RESULT):
-                    code = run_check_constraints(args, str(tmp_path))
-        assert code == 1
-
-    def test_severity_min_defaults_to_warn_when_none(self, tmp_path):
-        args = SimpleNamespace(
-            severity_min=None,
-            constraint_path_filter="",
-            constraint_file="/f.yml",
-        )
-        safe_result = {"success": True, "verdict": "SAFE"}
-        with patch(_RESOLVE_OFMT, return_value="json"):
-            with patch(_EVAL_EXPLICIT, return_value=safe_result) as mock_eval:
-                with patch(_PRINT_RESULT):
-                    run_check_constraints(args, str(tmp_path))
-        called_kwargs = mock_eval.call_args.kwargs
-        assert called_kwargs["severity_min"] == "warn"
-
-    def test_print_result_called_with_result_and_format(self, tmp_path):
-        args = _ns()
-        safe_result = {"success": True, "verdict": "SAFE"}
-        with patch(_RESOLVE_OFMT, return_value="toon"):
-            with patch(_ASYNCIO_RUN, return_value=safe_result):
-                with patch(_PRINT_RESULT) as mock_print:
-                    run_check_constraints(args, str(tmp_path))
-        mock_print.assert_called_once_with(safe_result, "toon")
-
-    def test_path_filter_passed_to_evaluate_explicit(self, tmp_path):
-        args = _ns(constraint_file="/f.yml", constraint_path_filter="src/**")
-        safe_result = {"success": True, "verdict": "SAFE"}
-        with patch(_RESOLVE_OFMT, return_value="json"):
-            with patch(_EVAL_EXPLICIT, return_value=safe_result) as mock_eval:
-                with patch(_PRINT_RESULT):
-                    run_check_constraints(args, str(tmp_path))
-        called_kwargs = mock_eval.call_args.kwargs
-        assert called_kwargs["path_filter"] == "src/**"
-
-
-def test_read_only_option_forwards_persist_false(tmp_path):
-    args = _ns(constraints_read_only=True)
-    safe_result = {"success": True, "verdict": "SAFE"}
-    with patch(_RESOLVE_OFMT, return_value="json"):
-        with patch(_ASYNCIO_RUN, return_value=safe_result):
-            with patch(_PRINT_RESULT):
-                with patch(
-                    "tree_sitter_analyzer.cli.commands.constraint_check_command._run_tool"
-                ) as run_tool:
-
-                    async def result():
-                        return safe_result
-
-                    run_tool.return_value = result()
-                    run_check_constraints(args, str(tmp_path))
-    assert run_tool.call_args.kwargs["persist"] is False
